Make `Table<Account>::iter` generic
